feat(logging): training-time viz callback + verify_e2e_gate.py
Adds the missing piece from agent/logging_wandb_plan.md §9.3 #6:
- TrainTimeVizCallback wires in when LoggingConfig.viz.log_during_training
  is True, writing PNGs to <fold_dir>/viz/epoch_{n}/ on every Nth
  validation epoch. Capped at n_train_predictions_logged.
- visualize_predictions_for_fold now skips ckpt-mtime resolution when a
  live LightningModule is provided (the training-time path).
- logging_setup silences git/urllib3/httpx/wandb/timm/PIL/etc. at WARNING
  to keep <fold_dir>/run.log focused on endo.* records when level_file=DEBUG.
- experiments/e2e_testing.py uses precision="32-true" (matches quickeval's
  band-aid for the bf16 NaN sensitivity surfaced in CLAUDE.md).

scripts/verify_e2e_gate.py — automated §9.3 success-criteria checker.
Filters by run state to ignore failed runs (e.g. the bf16 NaN one), reads
W&B history with unfiltered scan_history (the keys=[...] filter only
returns rows with all keys present, which W&B's row-per-event layout
breaks).

tests/viz/test_sample_tp_fp_fn.py — caps at N, deterministic, empty-event
handling.

Also adds python-dotenv to dependencies (used by wandb_init for .env load).

+++ experiments/e2e_testing.py
@@ -45,7 +45,12 @@ experiment = ExperimentConfig(
         num_workers=4,
         base_lr=2e-4,
         warmup_epochs=0,
-        precision="bf16-mixed",
+        # 32-true mirrors `experiments/quickeval.py` — bf16-mixed has shown
+        # intermittent NaN under runs longer than the smoke test on this
+        # repo (see top-level CLAUDE.md "Precision selection"). The e2e
+        # gate's purpose is to validate the W&B + logging plumbing, not to
+        # debug precision instability.
+        precision="32-true",
         gradient_clip_val=1.0,
         log_every_n_steps=10,
     ),
